# trabaio/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API do Sistema de Petshop",
    "version": "1.0.0",
    "description": "API para gerenciamento de petshop com autenticação JWT"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Login de usuário",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string"},
                  "password": {"type": "string"}
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "token": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {"description": "Credenciais inválidas"},
          "404": {"description": "Usuário não encontrado"},
          "500": {"description": "Erro interno do servidor"}
        }
      }
    },
    "/auth/admin": {
      "get": {
        "summary": "Exibir formulário de login ADM",
        "tags": ["Autenticação - ADM"],
        "responses": {
          "200": {"description": "Formulário de login ADM"},
          "302": {"description": "Redireciona para links se já logado"}
        }
      },
      "post": {
        "summary": "Processar login ADM",
        "tags": ["Autenticação - ADM"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string"},
                  "password": {"type": "string"}
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Login ADM bem sucedido (redireciona)"},
          "400": {"description": "Requisição inválida"},
          "401": {"description": "Senha incorreta"},
          "404": {"description": "Usuário não encontrado"},
          "500": {"description": "Erro interno do servidor"}
        }
      }
    },
    "/auth/admin/api-links": {
      "get": {
        "summary": "Página de links da API para ADM",
        "tags": ["Autenticação - ADM"],
        "security": [
          {"AdminAuth": []}
        ],
        "responses": {
          "200": {"description": "Links da API ADM"},
          "302": {"description": "Redireciona para login ADM se não autenticado"}
        }
      }
    },
...
    "/pets": {
      "get": {
        "summary": "Listar todos os pets",
        "tags": ["Pets"],
        "security": [
          {"BearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Lista de pets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar novo pet",
        "tags": ["Pets"],
        "security": [
          {"BearerAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pet criado com sucesso"
          },
          "500": {
            "description": "Erro ao cadastrar o pet"
          }
        }
      }
    },
    "/pets/{id}": {
      "get": {
        "summary": "Buscar pet por ID",
        "tags": ["Pets"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do pet"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do pet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "404": {"description": "Pet não encontrado"},
          "500": {"description": "Erro ao buscar pet"}
        }
      },
      "put": {
        "summary": "Atualizar pet por ID (completo)",
        "tags": ["Pets"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do pet"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetCreation"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Pet atualizado com sucesso"},
          "404": {"description": "Pet não encontrado"},
          "500": {"description": "Erro ao atualizar o pet"}
        }
      },
      "patch": {
        "summary": "Atualizar pet por ID (parcial)",
        "tags": ["Pets"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do pet"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetCreation"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Pet atualizado parcialmente com sucesso"},
          "404": {"description": "Pet não encontrado"},
          "500": {"description": "Erro ao atualizar o pet parcialmente"},
          "delete": {
            "summary": "Deletar pet por ID",
            "tags": ["Pets"],
            "security": [
              {"BearerAuth": []}
            ],
            "parameters": [
              {
                "in": "path",
                "name": "id",
                "required": true,
                "schema": {"type": "integer"},
                "description": "ID do pet"
              }
            ],
            "responses": {
              "200": {"description": "Pet deletado com sucesso"},
              "404": {"description": "Pet não encontrado"},
              "500": {"description": "Erro ao deletar o pet"}
            }
          }
        }
      }
    },
...
  }
}